int battery level

+++ IoT_auth_simulator/simulator/config/settings.py
@@ -86,8 +86,8 @@ class DeviceConfig:
     """Parameters that characterise an IoT device."""
 
     # Battery level range (%)
-    battery_min: float = 5.0
-    battery_max: float = 100.0
+    battery_min: int = 5.0
+    battery_max: int = 100.0
 
     # PSK (pre-shared key) byte length
     psk_length_bytes: int = 32
